+++ .gitignore
@@ -102,6 +102,7 @@ celerybeat.pid
 *.sage.py
 
 # Environments
+s3ts_env
 .env
 .venv
 env/
