+++ README.md
@@ -2,12 +2,11 @@
 
 ## Environment / Setup
 
-
 ```bash
 git clone https://github.com/rcote98/s3ts.git   # clone the repo
 cd s3ts                                         # move in the folder
 python3 -m venv s3ts_env                        # create virtualenv
-source env/bin/activate                         # activate it
+source s3ts_env/bin/activate                         # activate it
 pip install -r requirements.txt                 # install dependencies
 python -m pip install -e .                      # install dev package
 ```
