+++ requirements.txt
@@ -1,62 +1,48 @@
-absl-py==1.3.0
 aiohttp==3.8.3
-aiosignal==1.2.0
+aiosignal==1.3.1
 async-timeout==4.0.2
 attrs==22.1.0
-cachetools==5.2.0
 certifi==2022.9.24
 charset-normalizer==2.1.1
-contourpy==1.0.5
-cycler==0.11.0
 Deprecated==1.2.13
-fonttools==4.37.4
-frozenlist==1.3.1
-fsspec==2022.8.2
-google-auth==2.12.0
-google-auth-oauthlib==0.4.6
-grpcio==1.49.1
+fire==0.4.0
+frozenlist==1.3.3
+fsspec==2022.11.0
 idna==3.4
 joblib==1.2.0
-kiwisolver==1.4.4
+lightning-utilities==0.3.0
 llvmlite==0.39.1
-Markdown==3.4.1
-MarkupSafe==2.1.1
-matplotlib==3.6.1
 multidict==6.0.2
-numba==0.56.2
+numba==0.56.4
 numpy==1.22.4
-oauthlib==3.2.1
+nvidia-cublas-cu11==11.10.3.66
+nvidia-cuda-nvrtc-cu11==11.7.99
+nvidia-cuda-runtime-cu11==11.7.99
+nvidia-cudnn-cu11==8.5.0.96
 packaging==21.3
-pandas==1.5.0
+pandas==1.5.2
 patsy==0.5.3
-Pillow==9.2.0
-protobuf==3.19.6
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-pyDeprecate==0.3.2
+Pillow==9.3.0
+protobuf==3.20.1
 pyparsing==3.0.9
 python-dateutil==2.8.2
-pytorch-lightning==1.7.7
-pytz==2022.4
+pytorch-lightning==1.8.3.post1
+pytz==2022.6
 PyYAML==6.0
 requests==2.28.1
-requests-oauthlib==1.3.1
-rsa==4.9
-scikit-learn==1.1.2
-scipy==1.8.1
+scikit-learn==1.1.3
+scipy==1.9.3
 six==1.16.0
-sktime==0.13.4
-statsmodels==0.13.2
-tensorboard==2.10.1
-tensorboard-data-server==0.6.1
-tensorboard-plugin-wit==1.8.1
+sktime==0.14.0
+statsmodels==0.13.5
+tensorboardX==2.5.1
+termcolor==2.1.1
 threadpoolctl==3.1.0
-torch==1.12.1
-torchmetrics==0.10.0
-torchvision==0.13.1
+torch==1.13.0
+torchmetrics==0.10.3
+torchvision==0.14.0
 tqdm==4.64.1
 typing_extensions==4.4.0
-urllib3==1.26.12
-Werkzeug==2.2.2
+urllib3==1.26.13
 wrapt==1.14.1
 yarl==1.8.1
