--- s3ts/datasets/generators,py
@@ -1,10 +1,0 @@
-"""
-    Automate sample generation for auxiliary tasks.
-"""
-
-
-from s3ts.datasets.processing import build_STSs
-
-import logging
-
-log = logging.Logger(__name__)
